Remove unnecessary string[] allocations (#72642)
The result of the `GetAllItems` call was a fresh `string[]` that was
being copied via collection initializer into a `List<T>`. That produces
unnecessary allocations. This is particularly problematic because these
allocations tend to end up in the LOH.

Changed the API to be more explicit and consumer to not do the
unnecessary allocations. This saved 73MB in LOH allocations.