Address unstable ordering in AbstractBuiltInCodeStyleDiagnosticAnalyzer
This class was previously non-deterministic about the value that it would assign to Descriptor based on it's ctor input. The process it previously used was to take the ImmutableDictionary it took into it's ctor and convert the keys from that dictionary into an immutable aray. From that, it would use the first item as the descrptor. This would lead to unstable values being used, as evidenced in https://github.com/dotnet/roslyn/issues/73070.

Additionally, as DiagnosticDescriptor's IEquatable didn't use CustomTags to differentiate between descriptors, I saw that several items being added to the dictionary by some callers were being duplicated to each other.

As part of this PR, I changed the AbstractBuiltInCodeStyleDiagnosticAnalyzer to no longer take in an ImmutableDictionary, but rather an ImmutableArray with tuples of the keys/values. This allows the caller to control a stable ordering of descriptors.