Populate the CurrentBlockAnalysisData's _reachingWrites when creating an OperationTreeAnalysis.
This is similar to what FlowGraphAnalysisData does upon inspecting a new block.

Inspection of the lightbulb speedometer test demonstrated that a significant portion of CPU (~8%) in our OOP was spent in AbstractRemoveUnusedParametersAndValuesDiagnosticAnalyzer. Looking at the code a bit indicated that the expensive calls to create/invoke the ControlFlowGraph weren't being short-circuited properly in the trivial cases I tried locally.

When a read operation is encountered, the code looks at all writes that are still unread. For each of these, it upates the map to indicate that the write has now been read. However, this wasn't working for parameters as the _reachingWrites collection didn't account for them.