Nullable annotate InMethodyBinder
This is in response to #73022 where there is a null reference exception
during binder. The cause is, very likely, a `null` symbol being passed
into the binder. It's not possible to fully root cause this at the
moment but adding asserts and NRT annotations to help us track it down
in the future.