Ignore drive casing when comparing on windows platforms (#73380)
Razor adds properties for razor files in msbuild targets. The casing for the section names can be inconsistent when pulling in the drive letter, causing the source generator not to run in hot reload scenarios. This normalizes the drive casing for section names and retrieval of options to fix hot reload for vs code