Remove semantic model keep-alive code in completion
There are quite a few providers that don't request/use the semantic model during their computations, and in the case where they are being used to commit / GetDescription, the model would be obtained/calculated unnecessarilly.

I don't believe the keep alive mechanism here is necessary anymore, as Solution.OnSemanticModelObtained caches the model when it's obtained for the active document in the solution.

Additionally, I switched a couple places in CommitManager to use JTF.Runn instead of WaitAndGetResult