Reduce allocations in PENamespaceSymbol.GetMembers() (#73794)
Reduce allocations in PENamespaceSymbol.GetMembers()

This method commonly shows up in profiles, and this change is very similar to an optimization made to make GetTypeMembers more efficient.

Just opening roslyn.sln and bringing up completion once in a file ends up hitting this codepath about 5,000 times in VS and about 50,000 times OOP. With these changes, I see about 60% of those requests in VS hit the cache, and about 95% of those OOP hit the cache. This is about worst case for hits in this cache too, as PENamespaceSymbol object seemed to be reused quite well.