LanguageServer: Suppress SymbolInformation obsoletion warnings
SymbolInformation has been marked as deprecated in the protocol
classes, per the current spec. The request handlers need to be
updated to use DocumentSymbol/WorkspaceSymbol but this is nontrivial -
they currently use a service to create instances of a special
subclass of SymbolInformation with VS-specific extensions, and
VS doesn't seem to have support for DocumentSymbol/WorkspaceSymbol.