Remove GetTempPath copy
This function was addedin 2016 when the compiler was still coming to
terms with its server process model. At that time we discovered that the
return for `Path.GetTempPath()` was different for different invocation
of our build task. This happened when more complex builds changed
environment variables that influenced temp directories.

Part of our reaction was adding this function that gave us more control
over how temporary paths could vary between builds. It let us have
explicit control over the working directory.

In the years since this change was made we've gotten a lot more mature
in our model here:

1. `Path.GetTemPath` is banned in the compiler via analyzers
2. The server process model is established and we are much more
   rigorous with how environment variables, temp paths, etc ... are
   handled.

As such this method serves very little purpose. It just makes our code
different and harder to understand. As such we're removing it.