Merge changes from a single DidChange notification
I'm currently investigating a customer profile where didChange notifications are showing very prevalently in CPU time. I've create this PR (https://github.com/dotnet/roslyn/pull/74267) to help with the underlying performance of creating the SourceText, however, there would still be potentially many source texts created for a single operation.

This PR attempts to alleviate that situation by changing the DidChange handler to create a single source text on the input TextDocumentContentChangeEvent array. This array can be quite large, such as when the user does a replace-all or multi-line edit in VS.

Roughly 50% of CPU time in the customer profile was spent in our DidChange handler (about 29 seconds of CPU time).