Remove double array allocation in SemanticTokens
This code was calling ArrayBuilder.ToArray, which would always allocate a new array as a result. Additionally, the arraybuilder itself would end up being large enough that it wouldn't end up going back into the pool. Thus, two large array were allocated on many calls into this method.

There really is no need for this method to use ArrayBuilder, as the code that ends up owning the resultant array doesn't attempt to put it back into a pool.

This shows up as about 0.5% of allocation in devenv in the customer profile I'm looking at, half of which should be removed by this change.