Reduce allocations in AnalyzerFileReference.GetAnalyzerTypeNameMap
This method showed up when investigating a recent regression in allocations due to a change in lightbulb behaviors. Although this method hasn't changed recently, it is one of the easiest ways to claw back some memory during this scenario.

Locally, opening the vs-platform solution with this change reduces allocations under GetAnalyzerTypeNameMap from about 1.9% (63 MB) to about 0.1% (5 MB)