Remove C++ sources from repo
The C++ sources are triggering CodeQL checks. Suppressing them / getting
exceptions is very tricky. Instead just removing this code. The majority
of it was written to validate our floating point parsing code compared
to the C++ impl. It's been ages since that was needed. Can recover it by
looking at this commit in the future if needed.