Extension type erasure (#74313)
# Erasure
The erasure is in `PEModuleBuilder.Translate`. 

Here are some cases where we need to erase:
- top-level: `E` => `Underlying`
- type arguments: `C<E>` => `C<Underlying>` (or similar cases with arrays and pointer types)

Extensions are erased by default, but in various scenarios we need to keep extensions:
1. When serializing the type for purpose of the extension erasure attribute, no extensions should be erased.
2. Extension types in containing type position should never be erased (`E.Nested`)

One caveat is that CCI translation is lazy. If you translate `E` with erasure, you get `Underlying`. But if you translate `C<E>` with erasure, we'll just return `C<E>`. The nested extension type will be erased at a later step (when `GetGenericArguments` is called). As a result, we rely on context to carry the information of whether further translations should erase or keep extensions. 

Note: because we ran into some limitations with this approach, we're planning to encode the "keepExtensions" flag into CCI adapters instead. This is described in test `TypeReference_Modopt_ExtensionAsTypeArgument` but will be a later PR.

# Synthesizing the erasure attribute
In `Compilation_WellKnownMembers.cs`, we add `SynthesizeExtensionErasureAttribute` which synthesizes an attribute with a TypedConstant of type `string` but holding a `TypeSymbol` value.
`SynthesizeExtensionErasureAttribute` is used in `MethodSymbol.cs` (for return type) and in `SourceParameterSymbolBase.cs` (for parameters).

The attribute is only synthesized when there is an extension that needs erasure. This is handled by `TypeSymbolExtensions.ContainsErasedExtensionType` and some updates to the `TypeSymbolExtensions.VisitType` infrastructure (we're adding the `asContainer` parameter and switching to use named delegate types `TypePredicate` and `TypeWithAnnotationsPredicate`). 

That `TypedConstant` is converted to a `MetadataSerializedType` in `AttributeDataAdapter.CreateSerializedType` (similar handling to `typeof` constants, which are converted to `MetadataTypeOf`).
We use that `MetadataSerializedType` in two places:
1. in `MetadataWriter.SerializeMetadataExpression`, to be serialized into a string
2. in `MetadataVisitor.Visit(MetadataSerializedType)` to track references/dependencies (part of the `MetadataVisitor`/`ReferenceIndexerBase`/`ReferenceIndexer` machinery).

## Serializing the type to a string
The type serialization is handled by `TypeNameSerializer` and it is extended to allow encoding references to type parameters. We add the `allowTypeParameters` parameter, which is only used for encoding extension erasure attributes.

## Keeping track of references from the attribute
`ReferenceIndexerBase` needs to index the references both for the underlying type and the unerased type (used in attribute with typeof-like format). So `ReferenceIndexerBase.VisitTypeReference` is extended to track both of those separately (see change to `IReferenceOrISignature`).
For example, if the type is `C<E>`, we'll visit the `C<E>` type reference as part of the emitted type (`C<Underlying>`) and as part of the erasure attribute (as `C<E>` unerased).

# Deserializing from string back to type

The type deserialization is handled by `TypeNameDecoder` and it is also extended to allow decoding references to type parameters. We add the `allowTypeParameters` parameter, which is only used for decoding extension erasure attributes.
In `TypeNameDecoder` we therefore need the ability to lookup type parameters by their index given the current context (current method/type). This capability is pulled up in the type hierarchy (see `TypeNameDecoder.GetGenericTypeParamSymbol` and `TypeNameDecoder.GetGenericMethodTypeParamSymbol`).

The actual decoding from `string` to `TypeSymbol` is in `PEParameterSymbol` which calls `PEModuleSymbol.TryDecodeExtensionErasureAttribute`.

# Explicit usage of the attribute

It is disallowed

# Unbounded erasures

Some erasures would be unbounded. For example, `extension E for C<E>`. We add detection for that for source scenarios.
PROTOTYPE comments capture follow-ups for PE/retargeting scenarios.