Reduce the number of sourcetexts fully recreated from the tree during brace completion enter
In the speedometer trace I'm looking at, 2.3% of allocations during the typing phase of the test are due to recreating SourceText from trees during typing an enter. This also accounts for 2.5% of CPU during that period, all of which is on the UI thread.

Instead, this code just creates the new SourceText from the old SourceText and the set of changes between them.