Reduce allocations during SourceText serialization
These allocations are showing up during the typing portion of the scrolling speedometer test as about 9.4% of allocations for that scenario.

The issue here is that in .Net framework, the BinaryWriter.Write method allocates a local array for the results of the Encoding.GetBytes call. .NET core does not allocate in this case, opting to call the GetBytes version that writes to an array given it (and uses a pooled array as such).

The change is to simulate what .NET core does, which should get rid of all these temporary array allocations when serializing source texts from .net fx.