Modify SyntaxNodeExtensions.GetMembers to instead execute a given lambda over the collection.
The scrolling speedometer profile I'm looking at is showing 0.7% of allocations during it's typing phase as boxing of the struct enumerators previously returned from SyntaxNodeExtensions.GetMembers.

Instead, as there is only one caller of this method (CSharpSyntaxFacts.AppendMembers), I've changed the method to take in a lambda and have it do the enumeration over the members, invoking the callback on each.