Reduce allocations in AbstractSymbolCompletionProvider.CreateItems
The linq expression shows up as about 0.3% of allocations in the profile I'm looking for in the typing section of the scrolling speedometer test.

MultiDictionary has a struct-based value that is great for the case where the valueset typically is only a single object (which is the case for this code).