Remove most usage of "new ArrayBuilder"
Most callers of this were intending to instead use ArrayBuilder.GetInstance (as that actually is what pulls items from the pool). Otherwise, the code would create a new ArrayBuilder (and consequently the ImmutableArray.Builder), reallocate until the array hit the appropriate size, and likely return the item to a pool that would never be pulled from.