Modify MergedNamespaceDeclaration.MakeChildrento to allocate less
This method is showing up in a profile via the ArrayBuilder.ToDictionary calls.

1) Two dictionaries are created each call. The second Dictionary is only used for enumeration, and thus we can avoid it.
2) The ToDictionary implementation cannot use a PooledDictionary, whereas our invocation for types can.
3) The ToDictionary creates ArrayBuilders for each unique key. This is bad if there are a large number of unique keys (such as for our invocation for types). By creating such a large number of ArrayBuilders, we lose the benefit of the pooling, as the vast majority of the time we won't be able to return or get builders from the pool.

I separated the processing of the namespaces and types collections into their own methods, as they have slightly different characteristics. Specifically:

1) The namespaces collection has few unique keys, but potentially large numbers of items in their value collections.
2) The types collection has many unique keys, and usually small numbers of items in their value collections.

Handling namespaces is pretty straight forward using a  Dictionary<string, ArrayBuilder<SingleNamespaceDeclaration>>. However, handling types required a bit less convention to get around the issue of creating a large number of ArrayBuilder instances. Instead, we use an object as the value, with it either being an ArrayBuilder<SingleTypeDeclaration> when an identity has been seen multiple times, or just a SingleTypeDeclaration when the identity has only been seen once.

MergedNamespaceDeclaration.MakeChildren is showing up as about 1.5% of allocations in the typing section of the speedometer scrolling perf test profile. These changes should shave that down to about 0.5% or so.