Modify ISyntaxTreeFactoryService.CreateSyntaxTree to take in an optional SourceText. (#74707)
* Modity ISyntaxTreeFactoryService.CreateSyntaxTree to take in an optional SourceText.

I've noticed debugging locally that for every keystroke I'm seeing two StringTexts created that span the entire buffer. The creation of these texts originates from ParsedSyntaxTree needing to create SourceText, and the only way it knowing to do so is by serializing the tree into a string and creating a StringText from it.

Modify DocumentState.TryReuseSiblingRoot to utilize the sibling's sourcetext in addition to it's tree.