Remove dependency on deprecated ref package
This removes our dependency on the deprecated Microsoft.NET.Build.Extensions
package. This is used under the hood when building a `net461` (or
earlier) desktop TFM against `netstandard2.0` references. It contains a
number of facade assemblies that bridge the gaps in a few places.

There is only one test that dependend on this. I migrated it to a .NET
core scenario that should cover the same area.