Reduce allocations from newline information allocated by ChangedText.GetLinesCore (#74728)
* Reduce allocations from newline information allocated by ChangedText.GetLinesCore.

The typing scenario in the speedometer scrolling test shows this as 10% of allocations.

The general idea here is that ChangedText doesn't need to keep track of line information as the SourceText that it wraps has that information. The complexity that was in ChangedText around newline splitting now needs to sit in both CompositeText and SubText as they need to understand how to expose their line collections.