Remove MS.CA.Test.Resources.Proprietary PackageReference (#75037)
* Remove MS.CA.Test.Resources.Proprietary

This package is the last remnants of the roslyn-internal repository. It
was still holding onto a few test assets that never got ported into the
open but were available through the NuPkg reference.

This package is becoming a problem with component governance as it
is a `netstandard1.3` package and hence brings in a lot of vulnerable
components.

Considered upgrading this package to `netstandard2.0` but that would
mean un-archiving the roslyn-internal repo, finding a place to put the
source code and setting up a new official pipeline for the package. That
is a pretty high price. Instead I decided to finish the work that
I should've completed a number of years ago.

Yes this does add a few more .dll into our git tree. These are files
that haven't changed in 10+ years so versioning won't be an issue. They
were all brought down already for restore hence it's not new files.