Fix TypeAttributes visibility checking
When checking visibility we need to use the VisibilityMask because TypeAttributes is a funky flags enum that has multiple entries with the same value (unlike normal flags enums).
See https://learn.microsoft.com/en-us/dotnet/api/system.reflection.typeattributes?view=net-8.0#examples

We got lucky because the IsPublic extension method isn't actually used anywhere so this was never hit.

The same applies to checking for interface, but TypeAttributes.ClassSemanticsMask is the same value as TypeAttributes.Interface so it happens to work, but it's still better to use the canonical code for clarity.