Fix the OOM error in scripting tests
Was able to reproduce the Scripting `OutOfMemoryException` issues on my
local machine. After looking at the test run in PerfView it was clear
that we were allocating a lot more memory than necessary and that was,
likey, the cause of the OOM that we saw. Given that it was running under
x86 this seems fairly likely.

The allocations were coming more from the ResultsProvider and IOperation
test assemlbies. But given that scripting is small our scheduling
algorithm often puts scripting into a slice that includes both these
test assemblies. Their allocations are what is pushing us to OOM. That
is also likely why we couldn't reproduce this by just running the
scripting tests locally.

This PR attempts to address the cases that I saw. There are a few cases
where I moved eager operations to lazy. The bigger change though is
moving a to using an interpolated string handler for
`RoslynDebug.Assert`. There are a number of cases where had roughly the
following code:

```csharp
Debug.Assert(condition, $"... {syntaxNode} ...");
```

On net472 that is going to eagerly allocate the message string which
realizes the entire syntax node. One assert alone was contributing 9GB
of allocations all by itself. Did a simple regex search for
`Debug.Assert` that had an interpolation and moved them in mass to
`RoslynDebug.Assert`. Now the strings are only realized if the assert
fails. Will discuss with the team if we want to think about an
analyzer here

PS: if the scripting tests OOM on this PR I may take tomorrow off.