Reduce allocations in the DkmClr layer
The impl of the DkmClr layer will recreate a `DkmClrType` for the same
`Type` many many times during a debugging unit test. That means the call
to `GetEvalAttributes`, which is quite expensive in terms of
allocations, is also repeated.

This implements a very simple cache so that we only execute
`GetEvalAttributes` once per `Type`.

This saves 200MB of allocations in the case where the unit test crashes
with OOM. It likely saves a lot more when the process doesn't OOM. After
this change it's no longer a significant source of allocations