Remove allocations in ErorrFacts
When profiling our unit tests I discovered that `ErrorFacts` are a
significant source of `string` allocations. In `Emit` for example it can
be over 200MB of allocations.

This is due to our use of `Enum.ToString` as well as concat that value
with well known constants. To avoid these allocations I changed our code
generator to produce these strings at compile time.

I was a bit concerned about the performance of the `switch` compared to
the existing code so I wrote a quick benchmark to demonstrate that there
is no perf lost here.

| Method             | Runtime              | Mean      | Ratio | Gen0    | Allocated | Alloc Ratio |
|------------------- |--------------------- |----------:|------:|--------:|----------:|------------:|
| EnumToString       | .NET 8.0             |  66.80 us |  1.00 |       - |   45046 B |        1.00 |
| EnumToStringHelper | .NET 8.0             |  15.76 us |  0.24 |       - |         - |        0.00 |
| EnumToString       | .NET Framework 4.7.2 | 512.30 us |  7.69 | 21.4844 |  140197 B |        3.11 |
| EnumToStringHelper | .NET Framework 4.7.2 |  17.70 us |  0.27 |       - |         - |        0.00 |