Reduce allocations in TypeSymbolExtensions.IsAtLeastAsVisibleAs (#75401)
* Reduce allocations in TypeSymbolExtensions.IsAtLeastAsVisibleAs

This method showed up in multiple profiles as about 0.6% of allocations in the Roslyn CA process in the csharp editing speedometer test.

This simply avoids those allocations by switching to a static lambda using the arg data to pass in a pooled object. Note that the arg passed into VisitType needed to be a class to allow the ref parameter to change it's value.