Fixing the OOM in scripting (#75372)
Dug into the Scripting OOM tests more and figured out what was going on. The scripting tests predominantly create `PortableExecutableReference` instances using `Stream` as the source. Under the hood this creates a lot of native allocations in the form of memory mapped files and `AllocHGlobal` memory. As `PortableExecutableReference` is not disposable, these allocations are cleaned up passively in finalizers.

This means that our tests rely on the finalizer running fast enough that it frees up enough native memory for the next scripting test that allocates native memory to succeed. When the scripting tests are run in isolation or in an x64 process this works out just fine. But when we run the tests on x86 and in combination with other test assemblies this can break down and lead to OOM situations.

This fix approaches the problem from three directions:

1. Scripting had an existing mechanism for intercepting files being converted to `PortableExecutableReference`. This change expands on that to make it more thoroughly plumbed through the scripting code base and to have a single place, `ScriptOptions`, where the intercept can be hooked up.
2. Creating a `ScriptTestBase` type which sets up the intercepting of file loads and then actively disposing the resulting `MetadataImageReference` on `Dispose`. This means the native memory is actively freed at the end of each test. It also provides a `ScriptOptions` value that derived tests can leverage.
3. Plumbing through `ScriptOptions` to the individual tests. 

After doing this I ran the script tests under PerfView in a number of configurations and verified that the memory was being actively freed and the heap size of the process was reduced as expected.


Co-authored-by: David Barbet <dibarbet@gmail.com>
Co-authored-by: Jan Jones <jan.jones.cz@gmail.com>
Co-authored-by: David Barbet <dibarbet@gmail.com>