Reduce lock usage during EditDistance.GetEditDistance and BKTree.Find (#75610)
* Reduce lock usage during EditDistance.GetEditDistance and BKTree.Find

The SimplePool and ArrayPool classes previously used always locked on Allocate/Free which is showing up in a trace David Kean referred us to. Instead, as the char arrays being allocated are typically quite small, utilize the stack for these temporary char array locations when possible, falling back to heap allocating in the rare occasions where it's not.