Fix issue in PatternMatcher code where it was using a caseInsensitive index (#75621)
Noticed this while looking at a pooling issue in related code. The issue here is that the code found a case sensitive index, but was still returning a value based on the case insensitive index in one of the code paths.

Also, made a slight perf optimization to start the case sensitive search at the already located case insensitive search index.