Reduce allocations in AbstractAsynchronousTaggerProvider.ComputeDifference
This method is showing up in a scrolling and typing profile I'm looking at. This method computes the difference between two TagSpanIntervalTrees given an updated snapshot.

To do so, it currently creates a new heap allocated TagSpan for each item in both given interval trees. However, it need not do that, as that TagSpan isn't passed out of this method. Instead, TagSpanIntervalTree now has an AddAllSpans method that populates to a collection of value tuples, which ComputeDifference can now use.